Move tests with dependencies to integration/external
