+++ tests/functional/round_trip/test_mirror.py
@@ -123,31 +123,6 @@ def test_ipynb_to_Rmd(nb_file, no_jupytext_version_number):
     assert_conversion_same_as_mirror(nb_file, "Rmd", "ipynb_to_Rmd")
 
 
-@pytest.mark.requires_pandoc
-@pytest.mark.parametrize(
-    "nb_file",
-    list_notebooks("ipynb", skip="(functional|Notebook with|flavors|invalid|305)"),
-)
-def test_ipynb_to_pandoc(nb_file, no_jupytext_version_number):
-    assert_conversion_same_as_mirror(nb_file, "md:pandoc", "ipynb_to_pandoc")
-
-
-@pytest.mark.requires_quarto
-@pytest.mark.parametrize(
-    "nb_file",
-    list_notebooks(
-        "ipynb",
-        skip="(functional|Notebook with|plotly_graphs|flavors|complex_metadata|"
-        "update83|raw_cell|_66|nteract|LaTeX|invalid|305|text_outputs|ir_notebook|jupyter|with_R_magic)",
-    ),
-)
-def test_ipynb_to_quarto(
-    nb_file,
-    no_jupytext_version_number,
-):
-    assert_conversion_same_as_mirror(nb_file, "qmd", "ipynb_to_quarto")
-
-
 @pytest.mark.requires_myst
 @pytest.mark.parametrize(
     "nb_file",
@@ -157,14 +132,6 @@ def test_ipynb_to_quarto(
 )
 def test_ipynb_to_myst(nb_file, no_jupytext_version_number):
     assert_conversion_same_as_mirror(nb_file, "md:myst", "ipynb_to_myst")
-
-
-@pytest.mark.requires_sphinx_gallery
-@pytest.mark.parametrize(
-    "nb_file", list_notebooks("ipynb_py", skip="(raw|hash|frozen|magic|html|164|long)")
-)
-def test_ipynb_to_python_sphinx(nb_file, no_jupytext_version_number):
-    assert_conversion_same_as_mirror(nb_file, "py:sphinx", "ipynb_to_sphinx")
 
 
 """---------------------------------------------------------------------------------
@@ -203,28 +170,6 @@ def test_spin_to_ipynb(nb_file, no_jupytext_version_number):
 @pytest.mark.parametrize("nb_file", list_notebooks("md"))
 def test_md_to_ipynb(nb_file, no_jupytext_version_number):
     assert_conversion_same_as_mirror(nb_file, "ipynb", "md_to_ipynb")
-
-
-@pytest.mark.parametrize("nb_file", list_notebooks("Rmd"))
-def test_Rmd_to_ipynb(nb_file, no_jupytext_version_number):
-    assert_conversion_same_as_mirror(nb_file, "ipynb", "Rmd_to_ipynb")
-
-
-@pytest.mark.requires_sphinx_gallery
-@pytest.mark.parametrize("nb_file", list_notebooks("sphinx"))
-def test_sphinx_to_ipynb(nb_file, no_jupytext_version_number):
-    assert_conversion_same_as_mirror(nb_file, "ipynb:sphinx", "sphinx_to_ipynb")
-
-
-@pytest.mark.requires_sphinx_gallery
-@pytest.mark.parametrize("nb_file", list_notebooks("sphinx"))
-def test_sphinx_md_to_ipynb(nb_file, no_jupytext_version_number):
-    assert_conversion_same_as_mirror(
-        nb_file,
-        {"extension": ".ipynb", "format_name": "sphinx", "rst2md": True},
-        "sphinx-rst2md_to_ipynb",
-        compare_notebook=True,
-    )
 
 
 """---------------------------------------------------------------------------------
