+++ tests/integration/cli/test_execute.py
@@ -1,9 +1,11 @@
 import shutil
+from io import StringIO
 from unittest import mock
 
 import pytest
+from nbformat.v4.nbbase import new_code_cell, new_notebook
 
-from jupytext import read
+from jupytext import read, reads, write
 from jupytext.cli import jupytext
 from jupytext.version import __version__
 
@@ -281,3 +283,41 @@ def test_cat_execute_does_not_update_the_metadata(sample_md_notebook, tmp_path):
     new_md_text = md_file.read_text()
     assert __version__ not in new_md_text
     assert "kernelspec" not in new_md_text
+
+
+@pytest.mark.requires_user_kernel_python3
+def test_skip_execution(tmpdir, cwd_tmpdir, tmp_repo, python_notebook, capsys):
+    write(
+        new_notebook(cells=[new_code_cell("1 + 1")], metadata=python_notebook.metadata),
+        "test.ipynb",
+    )
+    tmp_repo.index.add("test.ipynb")
+
+    jupytext(["--execute", "--pre-commit-mode", "test.ipynb"])
+    captured = capsys.readouterr()
+    assert "Executing notebook" in captured.out
+
+    nb = read("test.ipynb")
+    assert nb.cells[0].execution_count == 1
+
+    jupytext(["--execute", "--pre-commit-mode", "test.ipynb"])
+    captured = capsys.readouterr()
+    assert "skipped" in captured.out
+
+
+@pytest.mark.requires_user_kernel_python3
+@pytest.mark.skip_on_windows
+@pytest.mark.filterwarnings("ignore")
+def test_utf8_out_331(capsys, caplog):
+    py = "from IPython.core.display import HTML; HTML(u'\xd7')"
+
+    with mock.patch("sys.stdin", StringIO(py)):
+        jupytext(["--to", "ipynb", "--execute", "-"])
+
+    out, err = capsys.readouterr()
+
+    assert err == ""
+    nb = reads(out, "ipynb")
+    assert len(nb.cells) == 1
+    print(nb.cells[0].outputs)
+    assert nb.cells[0].outputs[0]["data"]["text/html"] == "\xd7"
